accelerators/voice-channel/server/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
gunicorn==23.0.0
websockets==13.1
httpx==0.27.2
azure-identity==1.19.0
python-dotenv==1.0.1
